requirements.txt
# TWE-RAG CPU-only dependencies (pinned versions)
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.0
pandas==2.2.2
networkx==3.3
rank-bm25==0.2.2
joblib==1.4.2
python-dateutil==2.9.0.post0
tqdm==4.66.4
streamlit==1.37.0
matplotlib==3.8.4
pytest==8.3.3
